Eshop/requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.3.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
decorator==5.1.1
Django==4.0.4
django-rest-knox==4.2.0
djangorestframework==3.13.1
dynaconf==3.1.8
hvac==0.11.2
idna==3.3
knox==0.1.14
kombu==5.2.4
loguru==0.6.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
prompt-toolkit==3.0.29
pycparser==2.21
pytz==2022.1
requests==2.27.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
validators==0.19.0
vine==5.0.0
wcwidth==0.2.5
